refactor: Move hooks into hooks directory
