+++ packages/react-responsive-pagination/src/hooks/useAvailableWidth/index.ts
