+++ packages/react-responsive-pagination/src/hooks/useAvailableWidth/useContentWidth.ts
