+++ packages/react-responsive-pagination/src/hooks/useAvailableWidth/useResizeNotifier.ts
