+++ packages/react-responsive-pagination/src/hooks/useFoutDetector.tsx
@@ -1,6 +1,6 @@
 import { useLayoutEffect } from 'react';
 import ResizeObserver from 'resize-observer-polyfill';
-import { getWidth } from '../../helpers/style';
+import { getWidth } from '../helpers/style';
 
 export function useFoutDetector(
   getElements: () => HTMLElement[] | null,
