+++ packages/react-responsive-pagination/src/hooks/useWidthCalculator/getViewMetrics.tsx
@@ -1,5 +1,5 @@
-import { objectZip } from '../../../helpers/object';
-import { getNonContentWidth, getElementWidth } from '../../../helpers/style';
+import { objectZip } from '../../helpers/object';
+import { getNonContentWidth, getElementWidth } from '../../helpers/style';
 
 export function getViewMetricsFromContainer<ItemKey extends string>(
   containerElement: HTMLElement,
