+++ packages/react-responsive-pagination/src/hooks/useWidthCalculator/useIsUnmounted.ts
