+++ packages/react-responsive-pagination/src/hooks/useWidthCalculator/widthCalculator/navWidthCalculator.ts
@@ -1,4 +1,4 @@
-import { NavType } from '../../../../view';
+import { NavType } from '../../../view';
 
 export function createNavWidthCalculator(widths: Params) {
   return function navWidthCalculator(type: NavType, enabled: boolean) {
