+++ packages/react-responsive-pagination/src/hooks/useWidthCalculator/widthCalculator/numberWidthCalculator.ts
