+++ packages/react-responsive-pagination/src/hooks/useWidthCalculator/widthCalculator/pageWidthCalculator.ts
